+++ src/Table.tsx
@@ -3,7 +3,7 @@ import { StatusBar } from 'expo-status-bar';
 import { ScrollView, StyleSheet, View } from 'react-native';
 import { useSaveSnapshot } from './persist';
 import { levelerMachine } from './machine';
-import { MachineContext } from './context';
+import { MachineContext } from './MachineContext';
 
 const styles = StyleSheet.create({
   container: {
